enc-feat: import ties

+++ src/importexport/encore/internal/parser/elem-text.h
@@ -60,8 +60,11 @@ struct EncLyric : EncMeasureElem {
 // TIE element: dir byte (+5) and startFlag (+6) encode arc direction. See ENCORE_FORMAT.md §6.7 Tie.
 struct EncTie : EncMeasureElem {
     bool isTieStart { false };      // true when dir byte has bit 7 or bit 1 set, or startFlag has bit 7 set
-    quint8 arcX1         { 0 };     // arc start x (element offset +10); only valid for size >= 18
-    quint8 arcX2         { 0 };     // arc end   x (element offset +12); only valid for size >= 18
+    // Arc endpoints, uint16. Reading them a byte at a time works only on a little-endian file:
+    // in a big-endian one the significant byte is the second, so a byte-wide read yields zero for
+    // every tie in the file. See ENCORE_FORMAT.md §6.7 Tie.
+    quint16 arcX1        { 0 };     // arc start x (element offset +10, or +8 before format 3.07)
+    quint16 arcX2        { 0 };     // arc end   x (element offset +12, or +10 before format 3.07)
     qint8 sourcePosition { -1 };    // staff position of source note (+14); -1 = all notes in chord
 
     using EncMeasureElem::EncMeasureElem;
